// types.ts
export type MovieSearchListType = {
    Title: string;
    Year: string;
    imdbID: string;
    Type: string;
    Poster: string;
};
export type infoForTableType = {
    [key: string]: string;
    Year: string;
    Released: string;
    Runtime: string;
    Genre: string;
    Director: string;
    Actors: string;
    Country: string;
    imdbRating: string;
};
export type MovieType = {
    Title: string;
    Plot: string;
    Poster: string;
};